modularizado y funcionando
