--- backend/.env
@@ -1,7 +1,0 @@
-PORT=3005
-
-DB_USER=postgres
-DB_NAME=go-job
-DB_PASSWORD=7532145t
-DB_HOST=localhost
-DB_PORT=5432
