--- backend/src/routes/job.js
@@ -1,4 +1,0 @@
-const router = Router();
-
-
-module.exports = router;
