--- backend/src/routes/service.js
@@ -1,4 +1,0 @@
-const router = Router();
-
-
-module.exports = router;
